Move key type determination tests into tests/
